+++ packages/flutterfire_cli/lib/src/commands/config.dart
@@ -54,6 +54,14 @@ class ConfigCommand extends FlutterFireCommand {
           'Skip the Y/n confirmation prompts and accept default options (such as detected platforms).',
     );
     argParser.addOption(
+      'platforms',
+      valueHelp: 'platforms',
+      mandatory: isCI,
+      help:
+          'Optionally specify the platforms to generate configuration options for '
+          'as a comma separated list. For example "android,ios,macos,web,linux,windows".',
+    );
+    argParser.addOption(
       'ios-bundle-id',
       valueHelp: 'bundleIdentifier',
       mandatory: isCI,
@@ -122,6 +130,26 @@ class ConfigCommand extends FlutterFireCommand {
 
   bool get yes {
     return argResults!['yes'] as bool || false;
+  }
+
+  List<String> get platforms {
+    final platformsString = argResults!['platforms'] as String?;
+    if (platformsString == null || platformsString.isEmpty) {
+      return <String>[];
+    }
+    return platformsString
+        .split(',')
+        .map((String platform) => platform.trim().toLowerCase())
+        .where(
+          (element) =>
+              element == 'ios' ||
+              element == 'android' ||
+              element == 'macos' ||
+              element == 'web' ||
+              element == 'linux' ||
+              element == 'windows',
+        )
+        .toList();
   }
 
   bool get applyGradlePlugins {
@@ -275,12 +303,30 @@ class ConfigCommand extends FlutterFireCommand {
 
   Map<String, bool> _selectPlatforms() {
     final selectedPlatforms = <String, bool>{
-      kAndroid: flutterApp!.android,
-      kIos: flutterApp!.ios,
-      kMacos: flutterApp!.macos,
-      kWeb: flutterApp!.web,
+      kAndroid: platforms.contains(kAndroid) ||
+          platforms.isEmpty && flutterApp!.android,
+      kIos: platforms.contains(kIos) || platforms.isEmpty && flutterApp!.ios,
+      kMacos:
+          platforms.contains(kMacos) || platforms.isEmpty && flutterApp!.macos,
+      kWeb: platforms.contains(kWeb) || platforms.isEmpty && flutterApp!.web,
+      if (flutterApp!.dependsOnPackage('firebase_core_desktop'))
+        kWindows: platforms.contains(kWindows) ||
+            platforms.isEmpty && flutterApp!.windows,
+      if (flutterApp!.dependsOnPackage('firebase_core_desktop'))
+        kLinux: platforms.contains(kLinux) ||
+            platforms.isEmpty && flutterApp!.linux,
     };
-    if (isCI || yes) {
+    if (platforms.isNotEmpty || isCI || yes) {
+      final selectedPlatformsString = selectedPlatforms.entries
+          .where((e) => e.value)
+          .map((e) => e.key)
+          .toList()
+          .join(',');
+      logger.stdout(
+        AnsiStyles.bold(
+          '${AnsiStyles.blue('i')} Selected platforms: ${AnsiStyles.green(selectedPlatformsString)}',
+        ),
+      );
       return selectedPlatforms;
     }
     final answers = promptMultiSelect(
@@ -351,6 +397,26 @@ class ConfigCommand extends FlutterFireCommand {
       );
     }
 
+    FirebaseOptions? windowsOptions;
+    if (selectedPlatforms[kWindows] != null && selectedPlatforms[kWindows]!) {
+      windowsOptions = await FirebaseWebOptions.forFlutterApp(
+        flutterApp!,
+        firebaseProjectId: selectedFirebaseProject.projectId,
+        firebaseAccount: accountEmail,
+        platform: kWindows,
+      );
+    }
+
+    FirebaseOptions? linuxOptions;
+    if (selectedPlatforms[kLinux] != null && selectedPlatforms[kLinux]!) {
+      linuxOptions = await FirebaseWebOptions.forFlutterApp(
+        flutterApp!,
+        firebaseProjectId: selectedFirebaseProject.projectId,
+        firebaseAccount: accountEmail,
+        platform: kLinux,
+      );
+    }
+
     final futures = <Future>[];
 
     final configFile = FirebaseConfigurationFile(
@@ -359,6 +425,8 @@ class ConfigCommand extends FlutterFireCommand {
       iosOptions: iosOptions,
       macosOptions: macosOptions,
       webOptions: webOptions,
+      windowsOptions: windowsOptions,
+      linuxOptions: linuxOptions,
       force: isCI || yes,
     );
     futures.add(configFile.write());
@@ -451,6 +519,8 @@ class ConfigCommand extends FlutterFireCommand {
           if (androidOptions != null) [kAndroid, androidOptions.appId],
           if (iosOptions != null) [kIos, iosOptions.appId],
           if (macosOptions != null) [kMacos, macosOptions.appId],
+          if (linuxOptions != null) [kLinux, linuxOptions.appId],
+          if (windowsOptions != null) [kWindows, windowsOptions.appId],
         ],
         paddingSize: 2,
       ),