+++ packages/flutterfire_cli/lib/version.g.dart
@@ -1,2 +1,2 @@
 // This file is generated. Do not manually edit.
-String cliVersion = '0.2.1+1';
+String cliVersion = '0.2.2+2';